+++ src/main/java/com/nhnacademy/frontend/admin/controller/BookCategoryController.java
@@ -1,8 +1,8 @@
 package com.nhnacademy.frontend.admin.controller;
 
-import com.nhnacademy.frontend.admin.domain.request.BookCategoryCreateRequestDto;
-import com.nhnacademy.frontend.admin.domain.request.BookCategoryUpdateRequestDto;
-import com.nhnacademy.frontend.admin.domain.response.BookCategoryResponseDto;
+import com.nhnacademy.frontend.common.adapter.dto.book.request.BookCategoryCreateRequestDto;
+import com.nhnacademy.frontend.common.adapter.dto.book.request.BookCategoryUpdateRequestDto;
+import com.nhnacademy.frontend.common.adapter.dto.book.response.BookCategoryResponseDto;
 import com.nhnacademy.frontend.admin.service.BookService;
 import com.nhnacademy.frontend.common.exception.ValidationFailedException;
 import jakarta.validation.Valid;
