+++ src/main/java/com/nhnacademy/frontend/admin/controller/BookTagController.java
@@ -1,7 +1,7 @@
 package com.nhnacademy.frontend.admin.controller;
 
-import com.nhnacademy.frontend.admin.domain.request.BookTagCreateRequestDto;
-import com.nhnacademy.frontend.admin.domain.response.BookTagResponseDto;
+import com.nhnacademy.frontend.common.adapter.dto.book.request.BookTagCreateRequestDto;
+import com.nhnacademy.frontend.common.adapter.dto.book.response.BookTagResponseDto;
 import com.nhnacademy.frontend.admin.service.BookService;
 import com.nhnacademy.frontend.common.exception.ValidationFailedException;
 import jakarta.validation.Valid;
