+++ src/main/java/com/nhnacademy/frontend/admin/service/BookServiceImpl.java
@@ -1,9 +1,9 @@
 package com.nhnacademy.frontend.admin.service;
 
 import com.nhnacademy.frontend.common.adapter.BookAdapter;
-import com.nhnacademy.frontend.admin.domain.request.*;
-import com.nhnacademy.frontend.admin.domain.response.*;
-import com.nhnacademy.frontend.book.domain.response.SimpleBookResponseDto;
+import com.nhnacademy.frontend.common.adapter.dto.book.response.SimpleBookResponseDto;
+import com.nhnacademy.frontend.common.adapter.dto.book.request.*;
+import com.nhnacademy.frontend.common.adapter.dto.book.response.*;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
