+++ src/main/java/com/nhnacademy/frontend/common/adapter/dto/book/request/BookCategoryMapCreateRequestDto.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.frontend.admin.domain.request;
+package com.nhnacademy.frontend.common.adapter.dto.book.request;
 
 import lombok.Data;
