+++ src/main/java/com/nhnacademy/frontend/common/adapter/dto/book/response/BookDetailResponseDto.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.frontend.admin.domain.response;
+package com.nhnacademy.frontend.common.adapter.dto.book.response;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
