+++ src/test/java/com/nhnacademy/frontend/book/service/BookServiceImplTest.java
@@ -1,9 +1,12 @@
 package com.nhnacademy.frontend.book.service;
 
 import com.nhnacademy.frontend.common.adapter.BookAdapter;
-import com.nhnacademy.frontend.admin.domain.request.*;
-import com.nhnacademy.frontend.admin.domain.response.*;
 import com.nhnacademy.frontend.admin.service.BookServiceImpl;
+import com.nhnacademy.frontend.common.adapter.dto.book.request.*;
+import com.nhnacademy.frontend.common.adapter.dto.book.response.BookCategoryMapResponseDto;
+import com.nhnacademy.frontend.common.adapter.dto.book.response.BookCategoryResponseDto;
+import com.nhnacademy.frontend.common.adapter.dto.book.response.BookTagMapResponseDto;
+import com.nhnacademy.frontend.common.adapter.dto.book.response.BookTagResponseDto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
